// MangoTech/Hospital/src/app/payment/payment/model/payment.model.ts
export class Payment {
        username: string;
        phone: string;
        detail: string;
        services: string;
        maindoctor: string;
        extradoctor: string;
        price: string;
        paid: string;
        unpaid: string;
        payments: string;
        date: string;
        unit: string;
        number: string;
        sale: string;
        percent: string;
        reason: string;
        expected: string;
        status: number;
        created: number;
}